requirements.txt
pyserial
crc